security_rabbit/requirements.txt
django-templated-mail==1.1.1
djoser==2.0.1
celery==3.1.24
django-celery-results==1.0.1
amqp==1.4.9
Faker==2.0.1
django-sslserver==0.21
Orange3==3.23.1
